firmware: allow replacements to be read at runtime
This is mainly for consumption in opnsense-patch to address a weakness
patching .in files.  Not my favourite way of duplicating things in
VERSIONFILE but it is the most direct way to access information we do
not have at runtime.  Adding another file just for this seems silly.

(cherry picked from commit 81e8c7dd431a2ec902c66d17d7bf2250b2ebceb5)