interfaces: fix media setting and a bit of consolidation
So ifconfig returns:

    media: autoselect (2500Base-T <full-duplex>)

Which throws off the check because while 2500Base-T is automatically
set it is not forced even though it is supposed to force "2500Base-T".

This started with FreeBSD 14 / OPNsense 24.7.  24.1 on FreeBSD 13 did not
exhibit this.  Though I'm surprised that nobody noticed until now.

Also found that "mediaopt" store was not properly aligned, but that is
more of a cosmetic problem and seems to have consistency issues that
could be driver specific.

Lastly, 'mediaopt' always implies 'media' so simplify the check.

(cherry picked from commit 24266ee1c894aaed43b1fec6b4d04ed91a28d521)