firmware: experiment to allow file-based replacements
When specifying a file with -R we go into "replace all"
mode for the things that the build would normally do
on ".in" suffixed files.  The quoting is wild but this
works:

  # cat src/bin/python3.link.in
  python%%CORE_PYTHON_DOT%%
  # src/sbin/opnsense-version -R src/bin/python3.link.in
  python3.11

This is to allow opnsense-patch to properly patch these
files instead of putting defunct placeholders into the
system.