firewall: also set groups for special IPv6 interfaces #9016
This has changed considerably in the last decade, but "stf" devices are
still being handled separately for valid reasons.  To get the right IPv6
device we need get_real_interface() which is anchored in interface code,
but has been shortened so many times now that we can argue we are simply
looking up a configuration so we move it to util.inc in order to avoid the
worst of adding the function call to filter.inc which may or may not be
aware of interfaces.inc depending on the caller.

To make this more flexible for the future add a new 'both' option to the
function to return an array of all addresses that it can give us so we can
iterate over it quickly.  This also allows to avoid code duplication in
callers.

FWIW, I'm really glad we worked on shortening get_real_interface() over the
years and now we're adding a bit more code to reduce a number of callers in
the near future because it is not uncommon to ask for all devices associated
to an interface (even though stf as in 6RD and 6TO4 is a fringe use case).