dnsmasq: Allow disabling local for DHCP domains. (#9062)
If dnsmasq is the first DNS server in line, and you do not want to use a unique DHCP domain, forcing the DHCP domain to be local breaks all upstream queries for it. This might be wanted if you forward from Unbound to dnsmasq since it fixes DNS loops, but if dnsmasq is your primary forwarding resolver, it knows what to forward and what to answer.