Firewall: Aliases - add "quick" mode in update_tables.py to skip table size comparison when running scheduled. closes https://github.com/opnsense/core/issues/9111
This commit adds --quick flag which skips aliases when not changed or expired and some content is stored in the pf table.
In order to improve the handling, we're splitting the task of gathering content and determining if we should update the alias, which saves memory and (certainly on small machines) processing time.

Some small readability changes are also implemented, moved some functionality to the alias object and used these in the update_tables.py script.