bootgrid: simplify custom grid command additions a bit
specifically, the only way to make this work was to capture all
events on a parent element, which isn't all that intuitive. Since
the buttons are still being detached and moved somewhere else,
it is possible for events bound directly after Bootgrid instatiation
to be lost (there is no element to bind to). Therefore, these
events need to be set either before
the Bootgrid exists, or after the 'load.rs.jquery.bootgrid' phase