firewall: small regression in https://github.com/opnsense/core/commit/c33d2a0f5403190c99b613d8ea3d3bde81ade2d8, keep pagination when navigation isn't rendered
While this is somewhat unintuitive, the default mode for UIBootgrid
is to always expect pagination request parameters. If navigation is
turned off, these are simply set to null so all results are returned.
The previous commit broke the data fetch for cases where the navigation
was turned off (i.e. local and remote auth for ipsec connections).

the firewall live log is an exception here, as this table is fully
managed with custom logic and therefore doesn't need to use the
default request parameters.