make: do not rely on git to reset the file
The approach works but creates an error using the tools.git
"make test" target.  It would be nicer to not operate on the
actual config.xml that is checked in, but that's something
for later as tests need more of their own path use (and tmp
directory).

(cherry picked from commit edc11581b76ce9745c69b64bd0ea7c85069c2dff)