interfaces: support link-local IPv6 mode; closes #9328
At the moment we don't transition out of this again as ifdisabled is
not changed afterwards for tunnel device reasons, but maybe this needs
to be made more clever in the future.

Shuffle the remaining $realifv6 idiosyncrasy to 6rd 6to4 spots in the
code and don't make link-local eligible for PPPoE or gateways at this
point.  Both could work under certain constraints but since we never
had this feature before going two steps ahead doesn't seem very sensible.