system: switch int/bool to string in gateway properties; closes #9356
This is a broader commit for what looks like the right thing to do
for all fields actually anchored in the model for which even BooleanField
returns a string representation of the value.  Most should be benign but
the ones in Gateway.php and SettingsController.php should fix API display
issues.