system: no longer back up DUID but add some compat glue to importer
One could argue that restoring from /conf should take precedence here
in the mid term future to still be able to deliver a DUID from a single
/conf source and it's not that vividly documented anyway.  Let's see
where this leads us.  For now this should appease all use cases and we
do ignore the stale /conf ones beginning with 26.1 that may linger.