Services: Intrusion Detection - increase maintainability of suricata.yaml file.
The initial idea was to use php-yaml, parse/merge defaults and overlay the model options, but this doesn't seem to be very feasible due to the structure of the file.
Next best seems to be to remove all comments and default disabled options to make our options more clear, adding the additional + to inline if statements "{% ... +%}" decreases the chance of accidental breakage.