unbound: overview: fix quick allow/blocklist actions
The quick fix here is to block & allow on every defined policy. Ideally
one should be able to select a policy where an entry should apply
to in the case of a block action, and map back to the policy in case
of an allow action. The latter isn't possible in the current construct
yet as it needs a slight adjustment to the data format.